// src/app/shared/models/models.ts
export interface pageSelection {
  skip: number;
  limit: number;
}

export interface apiResultFormat {
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  data: Array<any>;
  totalData: number;
}
// ...
export interface BookingView{
  //to add center id
  bookingId:number;
  bookedBy:string;
  bookedByMobile:string;
  bookingDate:string;
  bookingTime:string;
  currencyCode:string;
  bookingDiscount:string;
  bookingPrice:string;
  statusName:string;
  doctorName:string;
  majorName:string;
}
// ...
export interface appointmentList {
  // name: string;
  // consultingDoctor: string;
  // treatment: string;
  // degree: string;
  // mobile: string;
  // email: string;
  // date: number;
  // time: number;
  // img: string;
  bookingId:number;
  bookedBy:string;
  bookedByMobile:string;
  bookingDate:string;
  bookingTime:string;
  currencyCode:string;
  bookingDiscount:string;
  bookingPrice:string;
  statusName:string;
  doctorName:string;
  majorName:string;
}